requirements.txt
# EverNear Cloud Backend
# Let pipecat-ai manage its own transitive deps (anthropic, cartesia, deepgram, numpy, pydantic)
pipecat-ai[deepgram,anthropic,cartesia,silero]==0.0.52
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
supabase>=2.11.0
httpx>=0.27.0
pyjwt[crypto]>=2.9.0
python-dotenv>=1.0.1
pydantic-settings>=2.5.0
aioapns>=3.2
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-httpx>=0.32.0
